Adding Authentication and Updating Status

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const fs = require("fs");
 const path = require("path");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -33,11 +34,15 @@ exports.getPosts = (req, res, next) => {
 
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
+
+  //check for errors
   if (!errors.isEmpty()) {
     const error = new Error("Entered Data is incorrect!");
     error.status = 422;
     throw error;
   }
+
+  //check for file upload
   if (!req.file) {
     const error = new Error("No Image Provided!");
     error.statusCode = 422;
@@ -47,20 +52,31 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path.replace("\\", "/");
+  let creator;
+
+  //create a new post
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Gustavo Test" },
+    creator: req.userId,
   });
+
   post
     .save()
     .then((result) => {
-      console.log(result);
-
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -97,6 +113,7 @@ exports.updatePost = (req, res, next) => {
   const content = req.body.content;
   let imageUrl = req.body.image;
 
+  //check for errors
   if (!errors.isEmpty()) {
     const error = new Error("Entered Data is incorrect!");
     error.status = 422;
@@ -107,12 +124,14 @@ exports.updatePost = (req, res, next) => {
     imageUrl = req.file.path;
   }
 
+  //check for file upload
   if (!imageUrl) {
     const error = new Error("No File uploaded!");
     error.statusCode = 422;
     throw error;
   }
 
+  //Find the post
   Post.findById(postId)
     .then((post) => {
       if (!post) {
@@ -120,9 +139,19 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      //check if user is the creator
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
+
+      //change the post content
       post.title = title;
       post.imageUrl = imageUrl;
       post.content = content;
@@ -151,12 +180,24 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
-      //check loggedin user
+      //check if user is the creator
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log("result");
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({ message: "Post Deleted!" });
     })
     .catch((err) => {
